misc/HtmlTestSuite/results/test-0348.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>CSS Box Alignment Level 3: parsing justify-content with valid values</title>
<link href="https://drafts.csswg.org/css-align-3/#propdef-justify-content" rel="help"/>
<meta content="justify-content supports the full grammar 'normal | &lt;content-distribution&gt; | &lt;overflow-position&gt;? [ &lt;content-position&gt; | left | right ]'." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
</head>
<body>
<script>
test_valid_value("justify-content", "normal");

// &lt;content-distribution&gt; = space-between | space-around | space-evenly | stretch
test_valid_value("justify-content", "space-between");
test_valid_value("justify-content", "space-around");
test_valid_value("justify-content", "space-evenly");
test_valid_value("justify-content", "stretch");

// &lt;overflow-position&gt;? [ &lt;content-position&gt; | left | right ]
// &lt;overflow-position&gt; = unsafe | safe
// &lt;content-position&gt; = center | start | end | flex-start | flex-end
test_valid_value("justify-content", "center");
test_valid_value("justify-content", "start");
test_valid_value("justify-content", "end");
test_valid_value("justify-content", "flex-start");
test_valid_value("justify-content", "flex-end");
test_valid_value("justify-content", "unsafe end");
test_valid_value("justify-content", "safe flex-start");
test_valid_value("justify-content", "left");
test_valid_value("justify-content", "unsafe right");
</script>
</body>
</html>
